+++ pages/index.tsx
@@ -4,6 +4,7 @@ import type { NextPage } from 'next';
 import { useEffect, useState } from 'react';
 import { reaction, runInAction } from 'mobx';
 import AppView from '../components/AppView';
+import DesktopUpdateBanner from '../components/DesktopUpdateBanner';
 import { useStores } from '../store/store-container';
 import { isDesktop, isDevelopment, isPlayground } from '../store/util';
 
@@ -72,6 +73,7 @@ const Home: NextPage = () => {
       }}
     >
       <AppView />
+      <DesktopUpdateBanner />
     </Container>
   );